Thu, Jun 29, 2023, 4:49 PM +02:00

+++ week04/pset4/filter-less/helpers.c
@@ -7,11 +7,11 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int blue = image[i][j].rgbtBlue;
-            int green = image[i][j].rgbtGreen;
-            int red = image[i][j].rgbtRed;
-            int average = (red +  green + blue) / 3;
-            
+            uint8_t blue = image[i][j].rgbtBlue;
+            uint8_t green = image[i][j].rgbtGreen;
+            uint8_t red = image[i][j].rgbtRed;
+            uint8_t average = (red +  green + blue) / 3;
+
             image[i][j].rgbtBlue = average;
             image[i][j].rgbtGreen = average;
             image[i][j].rgbtRed = average;
